+++ Managers/SettingsMan.cpp
@@ -20,6 +20,7 @@ namespace RTE {
 
 		m_FlashOnBrainDamage = true;
 		m_BlipOnRevealUnseen = true;
+		m_UnheldItemsHUDDisplayRange = 25;
 		m_EndlessMetaGameMode = false;
 		m_EnableCrabBombs = false;
 		m_CrabBombThreshold = 42;
@@ -129,6 +130,8 @@ namespace RTE {
 			reader >> m_BlipOnRevealUnseen;
 		} else if (propName == "MaxUnheldItems") {
 			reader >> g_MovableMan.m_MaxDroppedItems;
+		} else if (propName == "UnheldItemsHUDDisplayRange") {
+			SetUnheldItemsHUDDisplayRange(std::stof(reader.ReadPropValue()));
 		} else if (propName == "SloMoThreshold") {
 			reader >> g_MovableMan.m_SloMoThreshold;
 		} else if (propName == "SloMoDurationMS") {
@@ -299,6 +302,7 @@ namespace RTE {
 		writer.NewPropertyWithValue("FlashOnBrainDamage", m_FlashOnBrainDamage);
 		writer.NewPropertyWithValue("BlipOnRevealUnseen", m_BlipOnRevealUnseen);
 		writer.NewPropertyWithValue("MaxUnheldItems", g_MovableMan.m_MaxDroppedItems);
+		writer.NewPropertyWithValue("UnheldItemsHUDDisplayRange", m_UnheldItemsHUDDisplayRange);
 		writer.NewPropertyWithValue("SloMoThreshold", g_MovableMan.m_SloMoThreshold);
 		writer.NewPropertyWithValue("SloMoDurationMS", g_MovableMan.m_SloMoDuration);
 		writer.NewPropertyWithValue("EndlessMetaGameMode", m_EndlessMetaGameMode);