+++ Menus/SettingsGameplayGUI.h
@@ -7,6 +7,8 @@ namespace RTE {
 	class GUICollectionBox;
 	class GUICheckbox;
 	class GUITextBox;
+	class GUISlider;
+	class GUILabel;
 	class GUIEvent;
 
 	/// <summary>
@@ -54,6 +56,8 @@ namespace RTE {
 		GUICheckbox *m_ShowEnemyHUDCheckbox;
 		GUITextBox *m_MaxUnheldItemsTextbox;
 		GUITextBox *m_CrabBombThresholdTextbox;
+		GUISlider *m_UnheldItemsHUDDisplayRangeSlider;
+		GUILabel *m_UnheldItemsHUDDisplayRangeLabel;
 
 #pragma region Gameplay Settings Handling
 		/// <summary>
@@ -65,6 +69,11 @@ namespace RTE {
 		/// Updates the CrabBombThreshold textbox to override any invalid input, applies the setting value and removes its focus.
 		/// </summary>
 		void UpdateCrabBombThresholdTextbox();
+
+		/// <summary>
+		/// Updates the UnheldItemsHUDDisplayRange setting and label according to the slider value.
+		/// </summary>
+		void UpdateUnheldItemsHUDDisplayRange();
 #pragma endregion
 
 		// Disallow the use of some implicit methods.