+++ CHANGELOG.md
@@ -213,7 +213,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - New `ACraft` INI and Lua (R/W) property `ScuttleOnDeath` which can be used to disable the automatic self-destruct sequence when the craft's health drops down to zero.
 
-- New gameplay setting `MAX_INSERT_NAME_HERE` that hides the HUD of stranded items at a set distance.
+- New `Settings.ini` property `UnheldItemsHUDDisplayRange = numPixels` that hides the HUD of stranded items at a set distance. Default is 500 (25 meters).
+	Value of -1 or anything below means all HUDs will be hidden and the only indication an item can be picked up will be on the Actor's HUD when standing on top of it.
+	Value of 0 means there is no range limit and all items on Scene will display the pick-up HUD.
+	Valid range values are 1-1000, anything above will be considered as no range limit.
 
 </details>